restricting-access-with-permissions

+++ src/blog/api/views.py
@@ -1,7 +1,8 @@
 from msilib.schema import ServiceInstall
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 
 from account.models import Account
 from blog.models import BlogPost
@@ -9,6 +10,7 @@ from blog.api.serializers import BlogPostSerializer
 
 
 @api_view(['GET', ])
+@permission_classes((IsAuthenticated,))
 def api_detail_blog_view(request, slug):
 
     try:
@@ -22,12 +24,17 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT', ])
+@permission_classes((IsAuthenticated,))
 def api_update_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permissions to edit this post."})
 
     if request.method == 'PUT':
         serializer = BlogPostSerializer(blog_post, data=request.data)
@@ -40,12 +47,18 @@ def api_update_blog_view(request, slug):
 
 
 @api_view(['DELETE', ])
+@permission_classes((IsAuthenticated,))
 def api_delete_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        return Response(status=status.
+                        HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permissions to delete this post."})
 
     if request.method == "DELETE":
         operation = blog_post.delete()
@@ -58,9 +71,10 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST', ])
+@permission_classes((IsAuthenticated,))
 def api_create_blog_view(request):
 
-    account = Account.objects.get(pk=1)
+    account = request.user
 
     blog_post = BlogPost(author=account)
     if request.method == "POST":
